pyproject.toml
[tool.poetry]
name = "flask-crud"
version = "0.1.0"
description = ""
authors = ["ciaokitty <sharmasanvi125@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
watchdog = "^6.0.0"
pymongo = "^4.10.1"
flask-pymongo = "^2.3.0"
flask-smorest = "^0.45.0"
passlib = "^1.7.4"
flask-jwt-extended = "^4.7.1"
marshmallow = "^3.23.2"
email-validator = "^2.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
